Extract AudioFileProcessorView
Extract the class `AudioFileProcessorView` into its own files.

Cleanup includes.
